requirements.txt
torch==2.3.0
torchvision==0.18.0
streamlit==1.29.0
pillow==10.2.0
matplotlib==3.8.2
